First pass at codecov (coveralls)
I've decided to use cargo-travis (even though it feels a little bit like cheating to not roll a setup by hand first) because:
  - it handles kcov in an excellent way (grabs from source and builds when needed; uses target/ so that `cache: cargo` will effectively cache kcov too, etc.)
  - it supports coveralls out of the box
  - it seems extremely polished and most importantly: it saves us from having to maintain another thing

There aren't really any downsides specific to cargo-travis, just kcov and DWARF based code cov in general.
For one, we need sudo (though, if the config file in this commit works it should just be needed by the codecov deploy job).
kcov also has dependencies which matters because we're going to probably be building it from source (it is in the PPAs not the right versions).
And worst of all, it gives back some questionable coverage results.
But, it's what we've got for now so we'll make do.

N.B: cargo-travis also has a nice `doc-upload` subcommand that does exactly what the hackneyed other build stage does, but I'm going to refrain from switching to cargo-travis for that for now (one less thing to install - cache kind of negates this though).